requirements.txt
pathspec==0.5.2
GitPython>=2
